requirements.txt
# PatchCore and dependencies
torch
torchvision
timm
scikit-learn
tqdm
Pillow